qurt: Must use latest version of Eigen
The latest version of Eigen is required to build with the Hexagon
7.4 toolchain. Only certain C++11 features are supported and the
latest version if Eigen provides these tests. A patch is required
to add support for the Hexagon compiler.

These changes force the sync and then update of eigen and allow it
to be patched for qurt.

The eigen-3.2 submodule was removed as it is no longer needed by the
qurt build with the updated toolchain.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>